Books user relation

+++ src/books/books.service.ts
@@ -1,5 +1,6 @@
 import { Inject, Injectable } from '@nestjs/common/decorators';
 import {
+  ForbiddenException,
   ImATeapotException,
   NotFoundException,
 } from '@nestjs/common/exceptions';
@@ -14,39 +15,45 @@ export class BooksService {
     @Inject('IBooksRepository') private booksRepository: IBooksRepository,
   ) {}
 
-  async create({ title, pages }: CreateBookDto) {
+  async create(user_id: string, { title, pages }: CreateBookDto) {
     if (!pages) throw new ImATeapotException(errors.ZERO_PAGES);
 
-    const book = await this.booksRepository.create({ title, pages });
+    const book = await this.booksRepository.create(user_id, { title, pages });
     return book;
   }
 
-  async findAll() {
-    const books = await this.booksRepository.findAll();
+  async findAll(user_id: string) {
+    const books = await this.booksRepository.findAll(user_id);
     return books;
   }
 
-  async findOne(id: string) {
+  async findOne(user_id: string, id: string) {
     const book = await this.booksRepository.findOne(id);
 
     if (!book) throw new NotFoundException(errors.NON_EXISTENT_BOOK);
+    if (book.user_id !== user_id)
+      throw new ForbiddenException(errors.NOT_BELONG_TO_THE_USER);
 
     return book;
   }
 
-  async update(id: string, updateCrudDto: UpdateBookDto) {
+  async update(user_id: string, id: string, updateCrudDto: UpdateBookDto) {
     let book = await this.booksRepository.findOne(id);
 
     if (!book) throw new NotFoundException(errors.NON_EXISTENT_BOOK);
+    if (book.user_id !== user_id)
+      throw new ForbiddenException(errors.NOT_BELONG_TO_THE_USER);
 
     book = await this.booksRepository.update(id, updateCrudDto);
     return book;
   }
 
-  async delete(id: string) {
+  async delete(user_id: string, id: string) {
     const book = await this.booksRepository.findOne(id);
 
     if (!book) throw new NotFoundException(errors.NON_EXISTENT_BOOK);
+    if (book.user_id !== user_id)
+      throw new ForbiddenException(errors.NOT_BELONG_TO_THE_USER);
 
     return this.booksRepository.delete(id);
   }
